Move color utils from `common.ts` to `color-utils.ts`
