+++ src/ui/containers/candy-bar.ts
@@ -3,7 +3,7 @@ import { starterColors } from "#app/global-vars/starter-colors";
 import type { SpeciesId } from "#enums/species-id";
 import { TextStyle } from "#enums/text-style";
 import { addTextObject } from "#ui/text";
-import { rgbHexToRgba } from "#utils/common";
+import { rgbHexToRgba } from "#utils/color-utils";
 import { argbFromRgba } from "@material/material-color-utilities";
 
 export class CandyBar extends Phaser.GameObjects.Container {
